// src/enums/sciwheel/SciwheelReferenceType.ts
enum SciwheelReferenceType {
  F1000_ARTICLE = 'F1000_ARTICLE',
  RESEARCH_ARTICLE = 'RESEARCH_ARTICLE',
  EVIEW_ARTICLE = 'EVIEW_ARTICLE',
  BOOK = 'BOOK',
  CONFERENCE_PAPER = 'CONFERENCE_PAPER',
  WEBSITE = 'WEBSITE',
  SOFTWARE = 'SOFTWARE',
  THESIS = 'THESIS',
  PERSONAL_COMMUNICATION = 'PERSONAL_COMMUNICATION',
  PATENT = 'PATENT',
  REPORT_PAPER = 'REPORT_PAPER',
  LEGAL_CASE = 'LEGAL_CASE',
  NEWSPAPER_ARTICLE = 'NEWSPAPER_ARTICLE',
  MAGAZINE_ARTICLE = 'MAGAZINE_ARTICLE',
  PRESENTATION = 'PRESENTATION',
  FORUM_POST = 'FORUM_POST',
  BLOG_POST = 'BLOG_POST',
  MAP = 'MAP',
  FILM = 'FILM',
  BROADCAST = 'BROADCAST',
  OTHER = 'OTHER',
  MANUAL_ITEM_PDF = 'MANUAL_ITEM_PDF',
  MANUSCRIPT = 'MANUSCRIPT',
}

export default SciwheelReferenceType;