// src/enums/sciwheel/SciwheelIdType.ts
enum SciwheelIdType {
  PubMedId = 'PubMedId',
  PubMedCentralId = 'PubMedCentralId',
  DigitalObjectId = 'DigitalObjectId',
  LibraryItemId = 'LibraryItemId',
  PrimeArticleId = 'PrimeArticleId',
  CompositeArticleId = 'CompositeArticleId',
  PageTitle = 'PageTitle',
}

export default SciwheelIdType;